+++ android-pdf-viewer/src/main/java/com/github/barteksc/pdfviewer/annotation/core/shapes/PointUtil.kt
@@ -1,6 +1,7 @@
-package com.github.barteksc.pdfviewer.annotation.core
+package com.github.barteksc.pdfviewer.annotation.core.shapes
 
 import android.graphics.PointF
+import com.github.barteksc.pdfviewer.annotation.core.shapes.Edge
 
 /** Convert image coordinates to pdf coordinates and vice versa, no scale */
 fun PointF.convertCoordinatesFrom(pageHeight: Int) = PointF(x, pageHeight - y)
