+++ android-pdf-viewer/src/main/java/com/github/barteksc/pdfviewer/annotation/core/shapes/Shape.kt
@@ -1,4 +1,4 @@
-package com.github.barteksc.pdfviewer.annotation.core
+package com.github.barteksc.pdfviewer.annotation.core.shapes
 
 import android.graphics.PointF
